% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_delete.R
\name{hist_delete}
\alias{hist_delete}
\title{Delete unintended source excerpts}
\usage{
hist_delete(data, ID)
}
\arguments{
\item{data}{Your current source dataframe}

\item{ID}{Die text_ID der Quellenstellen, die man entfernen möchte}
}
\value{
The source dataframe ohne die Quellenstellen, die du löschen möchtest
}
\description{
Delete unintended source excerpts
}
\examples{
#Entferne die Quellenstellen mit den IDs 5 und 14
Xenophon_Corinthian <- Xenophon_Corinthian \%>\% hist_delete(ID = c(5, 14))
}
