% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_include.R
\name{hist_include}
\alias{hist_include}
\title{Neue Quellenstellen zu the_sources hinzufügen}
\usage{
hist_include(new_data)
}
\arguments{
\item{new_data}{Die Quellenstellen, die du hinzufügen möchtest als DataFrame}
}
\value{
Fertiges DataFrame inklusive der neuen Quellenstellen
}
\description{
Neue Quellenstellen zu the_sources hinzufügen
}
\examples{
# Das wird schwierig zu erstellen
}
