% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_sources.R
\name{hist_sources}
\alias{hist_sources}
\title{Erzeuge ein gefiltertes Quellen-DataFrame}
\usage{
hist_sources(
  author = NULL,
  context = NULL,
  event_type = NULL,
  event = NULL,
  work = NULL,
  citation = NULL,
  output_text = FALSE
)
}
\arguments{
\item{author}{Der Autor, nach dem die Quellen gefiltert werden sollen (z.B. "Xenophon")}

\item{context}{Der historische Kontext (z.B. "Corinthian War")}

\item{event_type}{Der Typ des Ereignisses (z.B. "battle")}

\item{event}{Das spezifische Ereignis (z.B. "Leuctra")}

\item{work}{Das Werk (z.B. "Hellenika")}

\item{citation}{Die Zitation (z.B. "1.1.1")}

\item{output_text}{Wenn \code{TRUE}, wird der Ergebnistext formatiert in der Konsole ausgegeben}
}
\value{
Ein gefiltertes DataFrame mit Quellenstellen (sofern \code{output_text = FALSE})
}
\description{
Diese Funktion filtert das interne Datenset \code{the_sources} nach den angegebenen Parametern.
Falls \code{output_text = TRUE}, wird der Ergebnistext formatiert als Ausgabe in der Konsole dargestellt.
}
\examples{
# Alle Quellenstellen von Xenophon über den Korinthischen Krieg als DataFrame
Xenophon_Corinthian <- hist_sources(author = "Xenophon", context = "Corinthian War")

# Als formatierter Text
hist_sources(author = "Xenophon", context = "Corinthian War", output_text = TRUE)
}
