% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_text.R
\name{hist_text}
\alias{hist_text}
\title{Output the source-paper}
\usage{
hist_text(filtered_sources)
}
\arguments{
\item{filtered_sources}{Automatically von hist_sources generierter Platzhaltername für das source dataframe}
}
\value{
The rendered text of the sources of the respective source dataframe (so the source-paper)
}
\description{
Output the source-paper
}
\examples{
# Example source dataframe created with hist_sources
Xenophon_Corinthian <- hist_sources(author="Xenophon",context="Corinthian War", output_text = FALSE)
# Outout the source-paper
Xenophon_Corinthian \%>\% hist_text()

}
