% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_sources.R
\name{hist_sources}
\alias{hist_sources}
\title{Create source dataframe nach Wunsch}
\usage{
hist_sources(
  author = NULL,
  context = NULL,
  event_type = NULL,
  event = NULL,
  work = NULL,
  citation = NULL,
  output_text = FALSE,
  source_data = the_sources
)
}
\arguments{
\item{author}{Der Autor nach die Quellen gefiltert werden sollen}

\item{context}{Der Kontext nach dem die Quellen gefiltert werden sollen}

\item{event_type}{Der Art des Ereignisses nach dem die Quellen gefiltert werden sollen}

\item{event}{Das Ereignis nach dem die Quellen gefiltert werden sollen}

\item{work}{Das Werk nach dem die Quellen gefiltert weden sollen}

\item{citation}{Die Zitation nach der die Quellen gefiltert werden sollen}

\item{output_text}{Möglichkeit sich den gerenderten Text der Quellen ausgeben zu lassen}
}
\value{
die gewünschten Quellenstellen als DataFrame
}
\description{
Create source dataframe nach Wunsch
}
\examples{
# Alle Quellenstellen von Xenophon über den Korinthischen Krieg als DataFrame
Xenophon_Corinthian <- hist_sources(author="Xenophon",context="Corinthian War", output_text=FALSE)
# Alle Quellenstellen von Xenophon über den Korinthischen Krieg als Text
Xenophon_Corinthian <- hist_sources(author="Xenophon",context="Corinthian War", output_text=TRUE)
}
